# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /manufacturing/orders:
    get:
      operationId: listProductionOrders
      tags: [manufacturing]
      summary: List production orders
      parameters:
        - $ref: "#/components/parameters/SearchParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [draft, released, in_progress, completed, cancelled]
        - name: priority
          in: query
          required: false
          schema:
            type: string
            enum: [low, medium, high]
      responses:
        "200":
          description: List of production orders
          content:
            application/json:
              schema:
                type: object
                required: [items, meta]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProductionOrder"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
    post:
      operationId: createProductionOrder
      tags: [manufacturing]
      summary: Create production order (auto-generates battery number + 8 stages)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductionOrderInput"
      responses:
        "201":
          description: Created order with stages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionOrderDetail"
  /manufacturing/orders/{id}:
    get:
      operationId: getProductionOrder
      tags: [manufacturing]
      summary: Get production order with stages, timeline, and genealogy
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Production order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionOrderDetail"
        "404":
          description: Not found
    patch:
      operationId: updateProductionOrder
      tags: [manufacturing]
      summary: Update production order header fields
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductionOrderUpdate"
      responses:
        "200":
          description: Updated production order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionOrderDetail"

  # ── Manufacturing — Stages ────────────────────────────────────────────────
...
  /logistics/dispatch-orders:
    get:
      operationId: listDispatchOrders
      tags: [logistics]
      summary: List dispatch orders
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [draft, confirmed, loaded, in_transit, delivered, cancelled] }
        - name: dealerId
          in: query
          schema: { type: string, format: uuid }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Dispatch orders list
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/DispatchOrder" }
                  total: { type: integer }
    post:
      operationId: createDispatchOrder
      tags: [logistics]
      summary: Create a dispatch order
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DispatchOrderInput" }
      responses:
        "201":
          description: Dispatch order created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DispatchOrder" }
  /logistics/dispatch-orders/{id}:
    get:
      operationId: getDispatchOrder
      tags: [logistics]
      summary: Get a dispatch order with batteries
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Dispatch order detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DispatchOrderDetail" }
    put:
      operationId: updateDispatchOrder
      tags: [logistics]
      summary: Update a dispatch order
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DispatchOrderInput" }
      responses:
        "200":
          description: Updated dispatch order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DispatchOrder" }
  /logistics/dispatch-orders/{id}/items:
    post:
      operationId: addDispatchItem
      tags: [logistics]
      summary: Add a battery to a dispatch order
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productionOrderId]
              properties:
                productionOrderId: { type: string, format: uuid }
      responses:
        "201":
          description: Item added
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DispatchItem" }
  /logistics/dispatch-orders/{id}/items/{itemId}:
    delete:
      operationId: removeDispatchItem
      tags: [logistics]
      summary: Remove a battery from a dispatch order
      parameters:
        - $ref: "#/components/parameters/IdParam"
        - name: itemId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Removed
  /logistics/dispatch-orders/{id}/status:
    post:
      operationId: advanceDispatchStatus
      tags: [logistics]
      summary: Advance dispatch order status
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status, actor]
              properties:
                status: { type: string, enum: [confirmed, loaded, in_transit, delivered, cancelled] }
                actor: { type: string }
                notes: { type: string }
      responses:
        "200":
          description: Updated dispatch order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DispatchOrder" }

  # ── Logistics — Packing Dashboard ─────────────────────────────────────────────
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SearchParam:
      name: search
      in: query
      required: false
      schema:
        type: string
...
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        default: 25
  schemas:
...
    PaginationMeta:
      type: object
      required: [total, page, pageSize, totalPages]
      properties:
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
        totalPages: { type: integer }
    StagePhoto:
      type: object
      required: [id, name, url, uploadedAt]
      properties:
        id: { type: string }
        name: { type: string }
        url: { type: string }
        uploadedAt: { type: string, format: date-time }
    OrderStage:
      type: object
      required: [id, productionOrderId, stageType, stageOrder, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        productionOrderId: { type: string, format: uuid }
        stageType:
          type: string
          enum: [cell_allocation, assembly, compression, bms_allocation, bms_programming, charging, testing, quality_control, packing]
        stageOrder: { type: integer }
        status:
          type: string
          enum: [pending, in_progress, paused, completed, approved, rejected]
        operatorName: { type: ["string", "null"] }
        supervisorName: { type: ["string", "null"] }
        startedAt: { type: ["string", "null"], format: date-time }
        pausedAt: { type: ["string", "null"], format: date-time }
        resumedAt: { type: ["string", "null"], format: date-time }
        completedAt: { type: ["string", "null"], format: date-time }
        approvedAt: { type: ["string", "null"], format: date-time }
        notes: { type: ["string", "null"] }
        photos:
          type: array
          items:
            $ref: "#/components/schemas/StagePhoto"
        stageData:
          type: object
          additionalProperties: true
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    ProductionOrder:
      type: object
      required: [id, orderNumber, batteryNumber, factoryManager, priority, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        orderNumber: { type: string }
        batteryNumber: { type: string }
        productId: { type: ["string", "null"], format: uuid }
        cellMatchId: { type: ["string", "null"], format: uuid }
        factoryManager: { type: string }
        currentStage:
          type: ["string", "null"]
          enum: [cell_allocation, assembly, compression, bms_allocation, bms_programming, charging, testing, quality_control, packing, "null"]
        priority:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [draft, released, in_progress, completed, cancelled]
        plannedStartDate: { type: ["string", "null"] }
        plannedEndDate: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    ProductionOrderDetail:
      allOf:
        - $ref: "#/components/schemas/ProductionOrder"
        - type: object
          properties:
            stages:
              type: array
              items:
                $ref: "#/components/schemas/OrderStage"
    ProductionOrderInput:
      type: object
      required: [factoryManager, priority]
      properties:
        productId: { type: ["string", "null"], format: uuid }
        factoryManager: { type: string }
        priority:
          type: string
          enum: [low, medium, high]
          default: medium
        plannedStartDate: { type: ["string", "null"] }
        plannedEndDate: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
    ProductionOrderUpdate:
      type: object
      properties:
        factoryManager: { type: string }
        priority:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [draft, released, in_progress, completed, cancelled]
        plannedStartDate: { type: ["string", "null"] }
        plannedEndDate: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }

    # ── Cell Grading Schemas ──────────────────────────────────────────────────
...
    Dealer:
      type: object
      required: [id, dealerCode, dealerName, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        dealerCode: { type: string }
        dealerName: { type: string }
        gstNumber: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        contactPerson: { type: ["string", "null"] }
        mobile: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        territory: { type: ["string", "null"] }
        creditLimit: { type: ["number", "null"] }
        status: { type: string, enum: [active, inactive] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    DispatchOrder:
      type: object
      required: [id, dispatchNumber, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        dispatchNumber: { type: string }
        dealerId: { type: ["string", "null"], format: uuid }
        customerName: { type: ["string", "null"] }
        transporter: { type: ["string", "null"] }
        vehicleNumber: { type: ["string", "null"] }
        driverName: { type: ["string", "null"] }
        driverMobile: { type: ["string", "null"] }
        dispatchDate: { type: ["string", "null"], format: date }
        status: { type: string, enum: [draft, confirmed, loaded, in_transit, delivered, cancelled] }
        notes: { type: ["string", "null"] }
        createdBy: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    DispatchOrderDetail:
      allOf:
        - $ref: "#/components/schemas/DispatchOrder"
        - type: object
          required: [items, shipmentEvents]
          properties:
            dealer:
              $ref: "#/components/schemas/Dealer"
            items:
              type: array
              items: { $ref: "#/components/schemas/DispatchItem" }
            shipmentEvents:
              type: array
              items: { $ref: "#/components/schemas/ShipmentEvent" }
    DispatchOrderInput:
      type: object
      properties:
        dealerId: { type: string, format: uuid }
        customerName: { type: string }
        transporter: { type: string }
        vehicleNumber: { type: string }
        driverName: { type: string }
        driverMobile: { type: string }
        dispatchDate: { type: string, format: date }
        notes: { type: string }
        createdBy: { type: string }
    DispatchItem:
      type: object
      required: [id, dispatchOrderId, productionOrderId, addedAt]
      properties:
        id: { type: string, format: uuid }
        dispatchOrderId: { type: string, format: uuid }
        productionOrderId: { type: string, format: uuid }
        batteryNumber: { type: string }
        orderNumber: { type: string }
        addedAt: { type: string, format: date-time }
    ShipmentEvent:
      type: object
      required: [id, dispatchOrderId, eventType, occurredAt]
      properties:
        id: { type: string, format: uuid }
        dispatchOrderId: { type: string, format: uuid }
        eventType: { type: string, enum: [ready_for_dispatch, loaded, in_transit, delivered, received_by_dealer] }
        actor: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        occurredAt: { type: string, format: date-time }
...